+++ pennylane_lightning/core/src/gates/Constant.hpp
@@ -85,7 +85,8 @@ using GateView = typename std::pair<GateOperation, std::string_view>;
     GateView{GateOperation::DoubleExcitationMinus, "DoubleExcitationMinus"},
     GateView{GateOperation::DoubleExcitationPlus, "DoubleExcitationPlus"},
     GateView{GateOperation::MultiRZ, "MultiRZ"},
-    GateView{GateOperation::GlobalPhase, "GlobalPhase"}};
+    GateView{GateOperation::GlobalPhase, "GlobalPhase"},
+    GateView{GateOperation::PSWAP, "PSWAP"}};
 
 using CGateView = typename std::pair<ControlledGateOperation, std::string_view>;
 [[maybe_unused]] constexpr std::array controlled_gate_names = {
@@ -118,6 +119,7 @@ using CGateView = typename std::pair<ControlledGateOperation, std::string_view>;
               "DoubleExcitationPlus"},
     CGateView{ControlledGateOperation::MultiRZ, "MultiRZ"},
     CGateView{ControlledGateOperation::GlobalPhase, "GlobalPhase"},
+    CGateView{ControlledGateOperation::PSWAP, "PSWAP"},
 };
 
 /**
@@ -151,6 +153,7 @@ using GeneratorView = typename std::pair<GeneratorOperation, std::string_view>;
     GeneratorView{GeneratorOperation::DoubleExcitationPlus,
                   "DoubleExcitationPlus"},
     GeneratorView{GeneratorOperation::GlobalPhase, "GlobalPhase"},
+    GeneratorView{GeneratorOperation::PSWAP, "PSWAP"},
 };
 
 using CGeneratorView =
@@ -178,6 +181,7 @@ using CGeneratorView =
                    "DoubleExcitationPlus"},
     CGeneratorView{ControlledGeneratorOperation::MultiRZ, "MultiRZ"},
     CGeneratorView{ControlledGeneratorOperation::GlobalPhase, "GlobalPhase"},
+    CGeneratorView{ControlledGeneratorOperation::PSWAP, "PSWAP"},
 };
 
 /**
@@ -238,6 +242,7 @@ using GateNWires = typename std::pair<GateOperation, std::size_t>;
     GateNWires{GateOperation::DoubleExcitationMinus, 4},
     GateNWires{GateOperation::DoubleExcitationPlus, 4},
     GateNWires{GateOperation::GlobalPhase, 1},
+    GateNWires{GateOperation::PSWAP, 2},
 };
 
 using CGateNWires = typename std::pair<ControlledGateOperation, std::size_t>;
@@ -266,6 +271,7 @@ using CGateNWires = typename std::pair<ControlledGateOperation, std::size_t>;
     CGateNWires{ControlledGateOperation::DoubleExcitationMinus, 4},
     CGateNWires{ControlledGateOperation::DoubleExcitationPlus, 4},
     CGateNWires{ControlledGateOperation::GlobalPhase, 1},
+    CGateNWires{ControlledGateOperation::PSWAP, 2},
 };
 
 /**
@@ -292,6 +298,7 @@ using GeneratorNWires = typename std::pair<GeneratorOperation, std::size_t>;
     GeneratorNWires{GeneratorOperation::DoubleExcitationMinus, 4},
     GeneratorNWires{GeneratorOperation::DoubleExcitationPlus, 4},
     GeneratorNWires{GeneratorOperation::GlobalPhase, 1},
+    GeneratorNWires{GeneratorOperation::PSWAP, 1},
 };
 
 using CGeneratorNWires =
@@ -312,6 +319,7 @@ using CGeneratorNWires =
     CGeneratorNWires{ControlledGeneratorOperation::DoubleExcitationMinus, 4},
     CGeneratorNWires{ControlledGeneratorOperation::DoubleExcitationPlus, 4},
     CGeneratorNWires{ControlledGeneratorOperation::GlobalPhase, 1},
+    CGeneratorNWires{ControlledGeneratorOperation::PSWAP, 1},
 };
 
 /**
@@ -355,6 +363,7 @@ using GateNParams = typename std::pair<GateOperation, std::size_t>;
     GateNParams{GateOperation::CSWAP, 0},
     GateNParams{GateOperation::MultiRZ, 1},
     GateNParams{GateOperation::GlobalPhase, 1},
+    GateNParams{GateOperation::PSWAP, 1},
 };
 
 /**
@@ -387,5 +396,6 @@ using CGateNParams = typename std::pair<ControlledGateOperation, std::size_t>;
     CGateNParams{ControlledGateOperation::DoubleExcitationPlus, 1},
     CGateNParams{ControlledGateOperation::MultiRZ, 1},
     CGateNParams{ControlledGateOperation::GlobalPhase, 1},
+    CGateNParams{ControlledGateOperation::PSWAP, 1},
 };
 } // namespace Pennylane::Gates::Constant